final adjustments - still needs to verify driving distance
@yuzhuz can you please analyze the best way to compute driving distance?

If we find out that the fields de.pickup and de.dropoff I'm using in rows 229-237 are good maybe we can use that approach.

the cl._outlier_deliveries table in BQ uses a different strategy, calculating what an outlier should be for each entity, but that also does not seem to solve the issue entirely as the data received from Hurrier looks bizarre sometimes.

I think we may need to talk to Igor and Pasquale to go to the bottom of this.